Remove TODO: inline isn't a workaround for no empty string rules
When I tried modifying the code to generate

    inline = $ => [
       ${inlineRules.map(x => '$.' + x).join(',\n')}
    ]

the grammar.js generation for Dhall got stuck in an infinite loop
trying to inline the 'whitespace' rule. With each iteration it
logged the error which I intended to work around.

    The rule `whitespace` matches the empty string.

The existing workaround of converting such rules into functions that
can be called from other non-empty string rules works well enough.